Automate release process (#8)
* Release using goreleaser

Use goreleaser to build and pass the version number as an ldflag

* Release project on merge to main branch

Use svu to create a new vMajor.Minor.Patch tag when a PR is merged to
main, then trigger goreleaser.  As the new tag has already been
created (and pushed) we then do not need to increment the version
number in .goreleaser.yml.